+++ projects/angular-material-extended/src/lib/matx-prompt/matx-prompt.models.ts
@@ -2,7 +2,7 @@ import { DialogPosition, DialogRole } from '@angular/material';
 import { Direction } from '@angular/cdk/bidi';
 import { ScrollStrategy } from '@angular/cdk/overlay';
 import { Observable } from 'rxjs';
-import { FormGroup, ValidatorFn, Validators } from '@angular/forms';
+import { Form, FormGroup, ValidatorFn, Validators } from '@angular/forms';
 
 export interface MatxPromptConfig extends MatxPromptData {
   /** ID for the dialog. If omitted, a unique one will be generated. */
@@ -78,7 +78,9 @@ export interface MatxPromptInput {
   noneText?: string;
   displayField?: string;
   displayWith?: (option) => string;
-  filterBy?: (form: FormGroup) => (value: string) => Observable<any[]>;
+  filterBy?: (form?: FormGroup) => (value: string) => Observable<any[]>;
+  disabled?: (form?: FormGroup) => void;
+  ngModelChange?: (value, form?: FormGroup) => void;
 }
 
 export interface MatxPromptAction {