+++ projects/angular-material-extended/src/lib/matx-select/matx-select.component.ts
@@ -40,6 +40,10 @@ export class MatxSelectComponent extends DefaultValueAccessor implements OnInit,
 
   @Input() floatLabel: 'auto' | 'always' | 'never';
 
+  @Input() set disabledControl(disabled: string | boolean) {
+    this.disabled = disabled;
+  }
+
   @Input() set disabled(disabled: string | boolean) {
     if (disabled === '' || disabled === true) {
       this.formControl.disable({emitEvent: false});