+++ projects/angular-material-extended/src/public_api.ts
@@ -16,3 +16,5 @@ export * from './lib/matx-select/matx-select.component';
 export * from './lib/matx-prompt/matx-prompt.module';
 export * from './lib/matx-prompt/matx-prompt.component';
 export * from './lib/matx-prompt/matx-prompt.controller';
+export * from './lib/matx-datepicker/matx-datepicker.component';
+export * from './lib/matx-datepicker/matx-datepicker.module';