+++ src/app/matx-datepicker-demo/matx-datepicker-demo.component.ts
@@ -1,0 +1,24 @@
+import { Component, OnInit } from '@angular/core';
+import { FormBuilder, FormGroup } from '@angular/forms';
+
+@Component({
+  selector: 'app-matx-input-demo',
+  templateUrl: './matx-datepicker-demo.component.html'
+})
+export class MatxDatepickerDemoComponent implements OnInit {
+
+  form1Model = {
+    matxDatepicker1: null,
+  };
+
+  form2: FormGroup;
+
+  constructor(private fb: FormBuilder) {}
+
+  ngOnInit(): void {
+    this.form2 = this.fb.group({
+      matxDatepicker2: null,
+    });
+  }
+
+}