+++ CHANGELOG.md
@@ -1,3 +1,97 @@
+# 0.6.0
+
+- Add `ngModelChange` and `disabled` attributes to `MatxPromptInput`
+- Add `disabledControl` input to `matx-input`, `matx-autocomplete`, `matx-select`, and `matx-gmap-select`
+- Add `matx-datepicker` control
+
+# 0.5.20
+
+- fix prompt close on backdrop click
+
+# 0.5.19
+
+- fix prompt button disabled when loading
+
+# 0.5.18
+
+- fix prompt loading buttons
+
+# 0.5.17
+
+- fix filterBy call with form of matx-prompt autocomplete
+
+# 0.5.16
+
+- add form to matx-prompt autocomplete filterBy
+
+# 0.5.15
+
+- add disabled, hideRequiredMarker and floatLabel inputs
+
+# 0.5.14
+
+- add type=button to matx-autocomplete buttons
+
+# 0.5.13
+
+- add matPrefix and matSuffix to matx-input
+
+# 0.5.12
+
+- fix matx-input number value, min, max, and step
+
+# 0.5.11
+
+- fix support numbers for matx-select and matx-prompt-input
+
+# 0.5.10
+
+- fix pre-selected string value in matx-select
+
+# 0.5.9
+
+- fix empty inputs and callback type in matx-prompt
+
+# 0.5.8
+
+- add validators to matx-prompt form
+
+# 0.5.7
+
+- change disabled of matx-select to use reactive form attribute
+
+# 0.5.6
+
+- add disabled to matx-select
+
+# 0.5.5
+
+- stop select initial value change
+
+# 0.5.4
+
+- fix select initial value
+
+
+# 0.5.3
+
+- fix autocomplete initial value
+- fix autocomplete displayWith function
+- add prompt initial values
+- add prompt displayWith function to autocomplete input
+
+# 0.5.2
+
+- fix prompt-autocomplete and prompt-select
+
+# 0.5.1
+
+- fix public api
+
+# 0.5.0
+
+- Add Select and Prompt
+
 # 0.1.0
 
 - Add`matx-errors`